requirements-test.txt
# Testing Dependencies for Evvie Time Tracker

# Core Testing Framework
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-flask==1.3.0
pytest-asyncio==0.23.8
pytest-xdist==3.6.1
pytest-benchmark==4.0.0

# Test Data Generation
factory-boy==3.3.0
faker==26.0.0

# Time/Date Mocking
freezegun==1.5.1

# End-to-End Testing
playwright==1.45.0
pytest-playwright==0.5.0

# Code Quality Tools
black==24.4.2
flake8==7.1.0
mypy==1.11.0
bandit==1.7.9

# Performance Testing
locust==2.29.1

# Additional Testing Utilities
responses==0.25.0
pytest-timeout==2.3.1
pytest-env==1.1.3